Refactored usage of network ports to UInt16 instead of Int

+++ Sources/SwiftOSCIONIO/Shared/OSCHandlerBlock.swift
@@ -11,5 +11,5 @@ public typealias OSCHandlerBlock = @Sendable (
     _ message: OSCMessage,
     _ timeTag: OSCTimeTag,
     _ host: String,
-    _ port: Int
+    _ port: UInt16
 ) -> Void
