+++ Sources/SwiftOSCIONIO/TCP/Shared/OSCTCPHandlerProtocol.swift
@@ -14,7 +14,7 @@ protocol _OSCTCPHandlerProtocol: _OSCHandlerProtocol {
 }
 
 extension _OSCTCPHandlerProtocol {
-    func _handle(receivedData data: Data, remoteHost: String, remotePort: Int) {
+    func _handle(receivedData data: Data, remoteHost: String, remotePort: UInt16) {
         // This routine must accommodate more than one consecutive packet contained in the data
         // which may happen when multiple packets are sent rapidly from a client.
